WIP commit in CSound, pushing conversion from dB relative to full scale down into the BASS and DSound -specific volume setting code.
This allows DSound to just scale the value (to hundredths of a dB) and push it directly into its underlying buffer volume property, without extra conversion to and from linear. In the next commit I'll try to migrate BASS from linear to log mode, so that it can avoid linear conversions too.