The default target loudness is now, after dataset analysis, -7.4 LUFS. Related default levels have been adjusted to suit it.
From the new comment in CConfigIni:

There exists a particular large, well-known, well-curated, and regularly-updated collection of content for use with Taiko no Tatsujin simulators. A statistical analysis was performed on the the integrated loudness and true peak loudness of the thousands of songs within this collection as of late August 2018.

The analysis allows us to select a target loudness which results in the smallest total amount of loudness adjustment applied to the songs of that collection. The selected target loudness should result in the least-noticeable average adjustment for the most users, assuming their collection is similar to the exemplar.

The target loudness which achieves this is -7.4 LUFS.