InputAdjustTime is now consistently and centrally applied to input event timestamps just before they are processed, and is applied to nothing else.
This means that all, and only, input is now being adjusted, instead of input being selectively adjusted and some non-input values possibly being adjusted when they shouldn't be. This should improve behaviour overall but may (and this is a good thing) start to help reveal any pre-existing issues with code that should (but for whatever reason doesn't) base its calculations on the adjusted time of each input event when that input event is processed. I'll be looking for those as I continue my work.